trs: split ForeignEnv out of the Interp (RunCore boot prep)
The foreign task family's runtime state — console/file handles
(FSlot tables, $ungetc pushback), the finished/fataled/stop
latches, quiet suppression, plusargs, timescale, and the $display
scratch buffer — moves into a ForeignEnv struct (foreign.rs) that
owns the design-independent arms of foreign_action/foreign_value
wholesale, gates included.  The Interp keeps a thin dispatcher:
fe.action()/fe.value() first, and what they decline is exactly the
design-coupled remainder ($dump* needs the VCD writer and the
instance tree; unknown names go to the BDPI table).

Behavior-identical by construction (arm bodies moved verbatim; the
13-test regress passes).  This is the boundary the RunCore boot
path needs: a compiled artifact's foreign bounces can be serviced
by a ForeignEnv alone, with the declined arms acting as lazy
materialization triggers for the full Interp.
